+++ internal/memo.go
@@ -1,8 +1,6 @@
 package internal
 
 import (
-	"github.com/apache/arrow/go/v16/arrow"
-	"reflect"
 	"sync"
 )
 
@@ -25,31 +23,4 @@ func (m *MemosT[K, V]) LoadOrStore(key K, generateObject func() (*V, error)) (*V
 		memo.Object, memo.err = generateObject()
 	})
 	return memo.Object, memo.err
-}
-
-type Codec func(structValue reflect.Value, arr arrow.Array, arrIdx int) error
-
-var codecNoOp Codec = func(structValue reflect.Value, arr arrow.Array, arrIdx int) error {
-	return nil
-}
-
-type CodecMemoT = MemosT[string, Codec]
-
-var CodecMemo = &CodecMemoT{}
-
-type NamespaceMemo struct {
-	// Root and non-root FQN as keys
-	ResolvedFieldNameToIndices map[string][]int
-	// Non-root FQN as keys only
-	StructFieldsSet map[string]bool
-}
-
-type NamespaceMemosT MemosT[reflect.Type, NamespaceMemo]
-
-var NamespaceMemos = &NamespaceMemosT{}
-
-func (m *NamespaceMemosT) Load(structType reflect.Type) (*NamespaceMemo, error) {
-	return (*MemosT[reflect.Type, NamespaceMemo])(m).LoadOrStore(structType, func() (*NamespaceMemo, error) {
-		return generateNamespaceMemo(structType, nil)
-	})
 }
